pyproject.toml
[project]
name = "formforge-api"
version = "0.1.0"
description = "Backend Service for FormForge"
authors = [
    { name = "Marko Samoilovic", email = "msamoilovic@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "sqlalchemy>=2.0.43",
    "psycopg2-binary>=2.9.10",
    "alembic>=1.16.5",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "google-generativeai>=0.8.5",
    "google-genai>=1.56.0",
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.20",
    "email-validator>=2.0.0",
]

[project.scripts]
dev = "scripts:run_dev"
seed = "scripts:seed_db"
reset-db = "scripts:reset_db"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
